school-management/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.0.1
certifi==2025.1.31
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
fonttools==4.56.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
numpy==2.2.2
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
pyasn1==0.6.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
threadpoolctl==3.5.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
